--- a/Makefile
+++ b/Makefile
@@ -2,7 +2,7 @@ all: gengetoptions
 
 help: ## display this help message
 	@echo "Please use \`make <target>' where <target> is one of"
-	@awk -F ':.*?## ' '/^[a-zA-Z]/ && NF==2 {printf "\033[36m  %-25s\033[0m %s\n", $$1, $$2}' $(MAKEFILE_LIST) | sort	
+	@awk -F ':.*?## ' '/^[a-zA-Z]/ && NF==2 {printf "\033[36m  %-25s\033[0m %s\n", $$1, $$2}' $(MAKEFILE_LIST) | sort
 
 # requires getoptions to be installed
 # installing getoptions locally: https://github.com/ko1nksm/getoptions/#installation
@@ -12,5 +12,12 @@ gengetoptions: ## use getoptions to generate argument parsing
 # requires shellspec to be installed
 # installing shellspec locally: https://github.com/shellspec/shellspec/#installation
 # note: the CI uses the latest stable version available in brew: `brew tap shellspec/shellspec`, `brew install shellspec`
-test: ## automated testing using shellspec
+
+unit_tests: ## automated testing using shellspec: fast unit tests without external dependency
+	shellspec $(shell find spec/ -type f ! -name pull_performance_spec.sh ! -name spec_helper.sh)
+
+performance_tests:  ## automated testing using shellspec: performance integration tests pulling from GitHub.com/openedx
+	shellspec spec/pull_performance_spec.sh
+
+test: ## automated testing using shellspec: run all tests
 	shellspec