--- a/spec/pull_spec.sh
+++ b/spec/pull_spec.sh
@@ -33,28 +33,37 @@ Describe 'Pull with directory param'
     echo "cp $@"
   }
 
+  cd() {
+      echo "cd $@"
+  }
+
+  rm() {
+      echo "rm $@"
+  }
 
   setup() { pull_directory="pull_directory:local_dir" pull_repository="pull_repository" pull_branch="pull_branch"; VERBOSE=1; }
   BeforeEach 'setup'
 
   It 'calls everything properly'
     When call pull_translations
-    The lines of output should equal 18
     The output should equal 'Creating a temporary Git repository to pull translations into "./translations_TEMP"...
-git init -b main
-git remote add -f origin https://github.com/pull_repository.git
+git clone --branch=pull_branch --filter=blob:none --no-checkout --depth=1 https://github.com/pull_repository.git translations_TEMP
+cd translations_TEMP
 Done.
 Setting git sparse-checkout rules...
-git sparse-checkout set !*
+git sparse-checkout set --no-cone !*
 git sparse-checkout add pull_directory/**
 Done.
 Pulling translation files from the repository...
-git pull origin pull_branch
+git checkout HEAD
+rm -rf .git
+cd ..
 Done.
 Copying translations from "./translations_TEMP/pull_directory" to "./local_dir"...
 cp -r ./translations_TEMP/pull_directory/* local_dir/
 Done.
 Removing temporary directory...
+rm -rf translations_TEMP
 Done.
 
 Translations pulled successfully!'
@@ -73,6 +82,13 @@ Describe 'Pull filters'
     true
   }
 
+  cd() {
+      echo "cd $@"
+  }
+
+  rm() {
+    true # Omit output
+  }
 
   setup() {
       pull_directory="pull_directory"
@@ -86,16 +102,16 @@ Describe 'Pull filters'
 
   It 'sets correct sparse-checkout rules'
     When call pull_translations
-    The lines of output should equal 10
-    The output should equal 'git init -b main
-git remote add -f origin https://github.com/pull_repository.git
-git sparse-checkout set !*
+    The output should equal 'git clone --branch=pull_branch --filter=blob:none --no-checkout --depth=1 https://github.com/pull_repository.git translations_TEMP
+cd translations_TEMP
+git sparse-checkout set --no-cone !*
 git sparse-checkout add pull_directory/**/ar/**
 git sparse-checkout add pull_directory/**/ar.*
 git sparse-checkout add pull_directory/**/es_419/**
 git sparse-checkout add pull_directory/**/es_419.*
 git sparse-checkout add pull_directory/**/fr_CA/**
 git sparse-checkout add pull_directory/**/fr_CA.*
-git pull origin pull_branch'
+git checkout HEAD
+cd ..'
   End
 End